Fix container module federation remote app name
There was an issue when trying to fetch the remote file from the
home remote app.
We changes the remote app name in the webpack file but we forgot to
change it in the container app when consuming it.

This commit fixes the issue by updating the container's app remote
name to homeApp.